requirements.txt
streamlit==1.31.1
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
datasets==2.17.0
huggingface-hub==0.20.3
wordcloud==1.9.3
PyYAML==6.0.1
openai==1.12.0
python-dotenv==1.0.1
plotly==5.18.0
kaleido